backend/pyproject.toml
[tool.poetry]
name = "test-fullstack"
version = "0.1.0"
description = ""
authors = ["caVenikk <sqqvt1@gmail.com>"]
readme = "README.md"
packages = [{include = "test_fullstack"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.1.1"
fastapi = "^0.103.1"
sqlalchemy = "^2.0.21"
alembic = "^1.12.0"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
